split the checkout route file into Sh oppingCart and ProductsInCart in addition to Checkout
