+++ src/components/ShoppingCart/ShoppingCart.module.css
@@ -1,0 +1,14 @@
+.productsSection {
+  gap: 1rem;
+
+  @media screen and (min-width: 768px) {
+    margin-bottom: 2rem;
+  }
+}
+
+.productsSection {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  width: 100%;
+}
